requirements-dev.txt
google-cloud-storage==2.7.0
pytest>=7.2.2
pytest-cov
python-dotenv
